# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Cette API a pour but d'être le back end du projet Creative Beats",
    "version": "1.0.0",
    "title": "Swagger Creative Beats",
    "contact": {
      "email": "jonathan.ramassamy@u-psud.fr"
    }
  },
...
  "paths": {
    "/challengeAPI/createChallenge": {
      "post": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Ajouter un challenge",
        "operationId": "createChallenge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Challenge"
              }
            }
          },
          "description": "Le challenge à créer ",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/challengeAPI/updateChallenge": {
      "post": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Mettre à jour un challenge",
        "operationId": "updateChallenge",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "Jwt Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Challenge"
              }
            }
          },
          "description": "Les valeurs du challenge à mettre à jour",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/challengeAPI/getNextChallenge": {
      "get": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Récupérer le prochain challenge",
        "description": "",
        "operationId": "getNextChallenge",
        "responses": {
          "200": {
            "description": "Un objet challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/challengeAPI/getCurrentChallenge": {
      "get": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Récupérer le challenge courant",
        "description": "",
        "operationId": "getCurrentChallenge",
        "responses": {
          "200": {
            "description": "Un objet challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/challengeAPI/addParticipant": {
      "post": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Ajouter un participant à un challenge",
        "description": "",
        "operationId": "addParticipant",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "Jwt Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string"
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  },
                  "soundCloudUrl": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error on update. | Music already exist. | Utilisateur non connecté.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/challengeAPI/updateMusicState": {
      "post": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Ajouter un challenge",
        "operationId": "updateMusicState",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challenge_id": {
                    "type": "string"
                  },
                  "music_id": {
                    "type": "string"
                  },
                  "newState": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "description": "Un objet contenant l'id du challenge, l'id de la musique et le nouvel état",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/challengeAPI/postVote": {
      "post": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Poster un vote sur un challenge",
        "operationId": "postVote",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "Jwt Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challenge_id": {
                    "type": "string"
                  },
                  "music_id": {
                    "type": "string"
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "description": "Un objet contenant l'id du challenge, l'id de la musique et l'utilisateur votant",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
...
    "/challengeAPI/getLastFinishedChallenge": {
      "get": {
        "tags": [
          "/challengeAPI/"
        ],
        "summary": "Récupérer le dernier challenge fini",
        "operationId": "getLastFinishedChallenge",
        "responses": {
          "200": {
            "description": "Un objet avec le dernier challenge et l'utilisateur gagnant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "$ref": "#/components/schemas/Challenge"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "_id",
          "email",
          "userAlias",
          "userTag",
          "userPwd",
          "userRole",
          "userPoints",
          "userDescription",
          "userLinks",
          "musics",
          "userVerified",
          "telCrypted"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userAlias": {
            "type": "string"
          },
          "userTag": {
            "type": "string"
          },
          "userPwd": {
            "type": "string",
            "format": "password"
          },
          "userRole": {
            "type": "string"
          },
          "userPoints": {
            "type": "string"
          },
          "userDescription": {
            "type": "string"
          },
          "userLinks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "musics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userVerified": {
            "type": "boolean"
          },
          "telCrypted": {
            "type": "string"
          }
        }
      },
...
      "Challenge": {
        "type": "object",
        "required": [
          "_id",
          "processCreatif",
          "debutChallenge",
          "finChallenge",
          "challengeArtwork",
          "challengeName",
          "challengePrices",
          "musics",
          "electeurs"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "processCreatif": {
            "type": "string",
            "format": "date-time"
          },
          "debutChallenge": {
            "type": "string",
            "format": "date-time"
          },
          "finChallenge": {
            "type": "string",
            "format": "date-time"
          },
          "challengeArtwork": {
            "type": "string"
          },
          "challengeName": {
            "type": "string"
          },
          "challengeWinner": {
            "$ref": "#/components/schemas/User"
          },
          "musics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MusicChallenge"
            }
          }
        }
      },
      "MusicChallenge": {
        "type": "object",
        "required": [
          "user",
          "musicUrl",
          "state",
          "createDate",
          "vote"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "musicUrl": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "vote": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}